requirements_scraper.txt
flask~=3.0.3
pandas~=2.2.2
gunicorn
gcsfs~=2024.3.1
dash~=2.16.1
plotly~=5.21.0
google-cloud-secret-manager
git+https://github.com/Tanguy9862/scraper-new-drug-approvals.git
